% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connection.R
\name{db_connection}
\alias{db_connection}
\title{Make a database connection}
\usage{
db_connection(server = "DFBCV9TWVASP001", database = "GFBioSQL")
}
\arguments{
\item{server}{The server name}

\item{database}{The database name}
}
\description{
Make a database connection
}
