% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_length.R
\name{get_age_length}
\alias{get_age_length}
\title{Get age and length data for a given species and survey}
\usage{
get_age_length(species_id, survey_id, filter_na = TRUE,
  con = db_connection())
}
\arguments{
\item{species_id}{The species number}

\item{survey_id}{The survey number}

\item{filter_na}{Remove NAs?}

\item{con}{A connection object to a database}
}
\value{
A data frame with ages, lengths, sexes, and dates
}
\description{
Get age and length data for a given species and survey
}
\examples{
\dontrun{
get_age_length(403, 16)
}

}
