% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{get_species_id}
\alias{get_species_id}
\title{Get a species ID given a species name}
\usage{
get_species_id(species_name = NULL, con = db_connection())
}
\arguments{
\item{species_name}{A species name}

\item{con}{A connection object to a database}
}
\description{
Get a species ID given a species name
}
\examples{
\dontrun{
get_species_id("sebastes borealis")
get_species_id("Sebastes borealis")
get_species_id("SebAstEs BORealis")
head(get_species_id())
}
}
