% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{get_survey_id}
\alias{get_survey_id}
\title{Get a survey ID given a survey name}
\usage{
get_survey_id(survey_name = NULL, con = db_connection())
}
\arguments{
\item{survey_name}{A survey name}

\item{con}{A connection object to a database}
}
\description{
Uses \code{\link[base]{grep}}
}
\examples{
get_survey_id("1989 Hecate Strait Multispecies")
get_survey_id("Hecate Strait Multispecies")
get_survey_id("hecate strait mULTI")
get_survey_id("haida")
head(get_survey_id())
}
