% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{get_survey_name}
\alias{get_survey_name}
\title{Get a survey name given a survey ID}
\usage{
get_survey_name(survey_id = NULL, con = db_connection())
}
\arguments{
\item{survey_id}{A survey ID code}

\item{con}{A connection object to a database}
}
\description{
Get a survey name given a survey ID
}
\examples{
\dontrun{
get_survey_name(2)
head(get_survey_name())
}
}
