% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vb.R
\name{sample_posterior_vb}
\alias{sample_posterior_vb}
\title{Sample from a von Bertalanffy curve posterior}
\usage{
sample_posterior_vb(model, n = 100)
}
\arguments{
\item{model}{A Stan model fit with \code{\link{fit_vb}}}

\item{n}{The number of posterior samples}
}
\description{
Sample from a von Bertalanffy curve posterior
}
\examples{
m <- fit_vb(short_age_length$age, short_age_length$length)
set.seed(1)
sample_posterior_vb(m, n = 5)
}
