% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_age_length.R
\docType{data}
\name{short_age_length}
\alias{short_age_length}
\title{Example age-length data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 758 rows and 3 columns.}
\usage{
short_age_length
}
\description{
Example fish age-length data. Originally extracted from GFBio for shortraker
rockfish with some random noise added to the age and length columns.
}
\keyword{datasets}
